+++ apps/frontend/src/modules/auth/mocks/handlers/auth.ts
@@ -1,6 +1,7 @@
 import { signInHandlers } from "./sign-in.handler.js";
 import { signUpHandlers } from "./sign-up.handler.js";
+import { meHandlers } from "./me.handler.js";
 
-const handlers = [...signUpHandlers, ...signInHandlers];
+const handlers = [...signUpHandlers, ...signInHandlers, ...meHandlers];
 
 export { handlers as authHandlers };