+++ packages/shared/src/libs/enums/api-path.enum.ts
@@ -2,7 +2,6 @@ const APIPath = {
 	AUTH: "/auth",
 	PLAN_DAYS: "/plan-days",
 	PLANS: "/plans",
-	QUIZ: "/quiz",
 	QUIZ_QUESTIONS: "/quiz-questions",
 	TASKS: "/tasks",
 	USERS: "/users",